fix(rsc): hydrate while streaming (#786)
Trying minimal repro of https://github.com/hi-ogawa/vite-plugins/pull/785.

Well, actually the issue might be my `injectRscScript`. Just switching to Devon's `rsc-html-stream` seems to fix the issue :open_mouth: 

Okay, `injectRscScript` is waiting for `</body></html>`, but that won't come when streaming is finished, so we have to start injecting a way before that. We already do `createBufferedTransformStream`, so it should be safe to start injecting when we see `<body>` start tag.

_todo_

- [x] test