Started navigation example
Copied the basic vite example and pasted in the navigation example. It does ssr but fails at hydration. The SceneRSCView can't find the Navigation Context. In dev mode, vite has 2 versions of every navigation react file - and if it uses the wrong one then the context instances don't match. This doesn't happen in build/preview mode